remove JWTUtils class
