+++ back/src/main/java/org/superdev/coddy/user/utils/SecurityUtils.java
@@ -1,5 +1,9 @@
 package org.superdev.coddy.user.utils;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
@@ -14,13 +18,52 @@ import java.util.Random;
  */
 public class SecurityUtils {
 
+    private static final String JWT_ALGORITHM = "AES";
+
     private static final int SALT_SIZE = 32;
     private static final int KEY_LENGTH = 512;
     private static final int ITERATIONS = 1000;
     private static final String ALGORITHM = "PBKDF2WithHmacSHA512";
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(SecurityUtils.class);
+
     private SecurityUtils() {
         // this is an utility class. You do not need to instantiate it.
+    }
+
+
+    /**
+     * define all fields in the jwt token payload :
+     * <p>
+     * {@code
+     * {
+     * "login": "jdoe",
+     * "firstname": "John",
+     * "nbf": 1479975670,
+     * "permissions": ["write","read"],
+     * "exp": 1479976870,
+     * "lastname": "Doe"
+     * }
+     * }
+     * </p>
+     */
+    public enum PayloadFields {
+        LOGIN("login"),
+        FIRST_NAME("firstname"),
+        LAST_NAME("lastname"),
+        PERMISSIONS("permissions"),
+        NOT_BEFORE("nbf"),
+        EXPIRATION("exp");
+
+        private final String name;
+
+        PayloadFields(final String name) {
+            this.name = name;
+        }
+
+        public String getName() {
+            return this.name;
+        }
     }
 
     /**
@@ -39,6 +82,7 @@ public class SecurityUtils {
             SecretKey key = secretKeyFactory.generateSecret(spec);
             return key.getEncoded();
         } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
+            LOGGER.debug(e.getMessage(),e);
             throw new SecurityException(e);
         } finally {
             // erase the password in the char array in order to not retrieve it in the java memory
@@ -54,6 +98,23 @@ public class SecurityUtils {
         byte[] salt = new byte[SecurityUtils.SALT_SIZE];
         r.nextBytes(salt);
         return salt;
+    }
+
+
+    /**
+     * @return a random key with 256 characters
+     */
+    public static byte[] generateSecret() {
+        KeyGenerator keyGenerator;
+        try {
+            keyGenerator = KeyGenerator.getInstance(JWT_ALGORITHM);
+        } catch (NoSuchAlgorithmException e) {
+            LOGGER.debug(e.getMessage(), e);
+            LOGGER.error("Intern algorithm to create secret key doesn't exist");
+            return new byte[]{};
+        }
+        keyGenerator.init(256);
+        return keyGenerator.generateKey().getEncoded();
     }
 
     /**
